pyproject.toml
[tool.poetry]
name = "fudly-bot"
version = "2.0.0"
description = "Telegram bot for selling discounted food (Too Good To Go alternative for Uzbekistan)"
authors = ["shaxa2505"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
aiogram = ">=3.0.0"
python-dotenv = ">=0.19.0"
aiohttp = ">=3.9.0"
requests = ">=2.31.0"
psycopg = {extras = ["binary"], version = ">=3.2"}
psycopg-pool = ">=3.2"
python-json-logger = ">=2.0.0"
cachetools = ">=5.3.0"
redis = ">=5.0.0"
sentry-sdk = ">=1.40.0"
Pillow = ">=10.0.0"
qrcode = ">=7.4.0"
google-generativeai = ">=0.3.0"
pydantic = ">=2.0.0"
alembic = ">=1.13.0"
sqlalchemy = ">=2.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
pytest-asyncio = "^0.21.0"
mypy = "^1.7.0"
black = "^23.12.0"
ruff = "^0.1.7"
pre-commit = "^3.5.0"
types-requests = "^2.31.0"

[tool.black]
line-length = 100
target-version = ['py310']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.ruff]
line-length = 100
target-version = "py310"
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "C",   # flake8-comprehensions
    "B",   # flake8-bugbear
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false  # Start with false, then enable
ignore_missing_imports = false
strict_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
check_untyped_defs = true

[[tool.mypy.overrides]]
module = "aiogram.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "psycopg.*"
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers --cov=app --cov=handlers --cov-report=term-missing --cov-report=html"
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.coverage.run]
source = ["app", "handlers"]
omit = [
    "*/tests/*",
    "*/__pycache__/*",
    "*/venv/*",
    "*/.venv/*",
]

[tool.coverage.report]
precision = 2
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
